refactor: split quizGameConfigs into per-group files
Move 14 quiz configs from one 324-line file into 4 focused files:
- configs/language.tsx  — spelling, opposites, english-words, world-languages
- configs/knowledge.tsx — riddles, capitals, instruments, sports-quiz, continents
- configs/social.tsx    — emotions, family, healthy-food
- configs/math.tsx      — fractions, shapes-3d

Added defineConfig<Q>() factory in configs/types.ts for type-safe
config authoring without explicit any annotations.
quizGameConfigs.tsx becomes a thin barrel; no changes to consumers.

Closes #89
